Restructure main entrypoint
